starting the server class
